Refactoring Model.cs file
Moved all models from "Model.cs" file to coressponding folder
