+++ BusSolOnDB/BusSolOnDB.csproj
@@ -60,7 +60,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BusFleet.cs" />
+    <Compile Include="Models\BusFleet.cs" />
     <Compile Include="Constans.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
@@ -68,7 +68,9 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
-    <Compile Include="Model.cs" />
+    <Compile Include="Models\Bus.cs" />
+    <Compile Include="Models\Move.cs" />
+    <Compile Include="Models\Transaction.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Form1.resx">
