+++ BusSolOnDB/Constans.cs
@@ -4,7 +4,7 @@
     {
         public const int MinutesInHour = 60;
         public const int HoursInDay = 24;
-        public static string GetTimeFromNimutes(int minutes)
+        public static string GetTimeFromMinutes(int minutes)
         {
             return minutes / MinutesInHour + ":" + minutes % MinutesInHour;
         }
